Add helmet (#119)
* add https://www.npmjs.com/package/helmet for security

* add https://www.npmjs.com/package/helmet for security

+++ src/app.js
@@ -8,8 +8,10 @@ const mongoClient = require('./lib/mongo');
 const {fhirServerConfig, mongoConfig} = require('./config');
 const Prometheus = require('./utils/prometheus.utils');
 const env = require('var');
+const helmet = require('helmet');
 
 const app = express();
+app.use(helmet());
 app.use(Prometheus.requestCounters);
 app.use(Prometheus.responseCounters);
 Prometheus.injectMetricsRoute(app);
@@ -59,39 +61,43 @@ app.get('/version', (req, res) => {
 });
 
 app.get('/clean', async (req, res) => {
-    console.info('Running clean');
+    if (!env.IS_PRODUCTION) {
+        console.info('Running clean');
 
-    // Connect to mongo and pass any options here
-    let [mongoError, client] = await asyncHandler(
-        mongoClient(mongoConfig.connection, mongoConfig.options)
-    );
-    if (mongoError) {
-        console.error(mongoError.message);
-        console.error(mongoConfig.connection);
-        client.close();
-        res.status(500).json({success: false, error: mongoError});
-    } else {
-        //create client by providing database name
-        const db = client.db(mongoConfig.db_name);
-        var collection_names = [];
-        // const collections = await db.listCollections().toArray();
+        // Connect to mongo and pass any options here
+        let [mongoError, client] = await asyncHandler(
+            mongoClient(mongoConfig.connection, mongoConfig.options)
+        );
+        if (mongoError) {
+            console.error(mongoError.message);
+            console.error(mongoConfig.connection);
+            client.close();
+            res.status(500).json({success: false, error: mongoError});
+        } else {
+            //create client by providing database name
+            const db = client.db(mongoConfig.db_name);
+            let collection_names = [];
+            // const collections = await db.listCollections().toArray();
 
-        await db.listCollections().forEach(collection => {
-            console.log(collection.name);
-            if (collection.name.indexOf('system.') === -1) {
-                collection_names.push(collection.name);
+            await db.listCollections().forEach(collection => {
+                console.log(collection.name);
+                if (collection.name.indexOf('system.') === -1) {
+                    collection_names.push(collection.name);
+                }
+            });
+
+            console.info('Collection_names:' + collection_names);
+            for (const collection_index in collection_names) {
+                const collection_name = collection_names[parseInt(collection_index)];
+                console.log(collection_name);
+                console.log(['Removing: ', await db.collection(collection_name).countDocuments({}), ' documents from ', collection_name].join(''));
+                await db.collection(collection_name).deleteMany({});
             }
-        });
-
-        console.info('Collection_names:' + collection_names);
-        for (const collection_index in collection_names) {
-            const collection_name = collection_names[parseInt(collection_index)];
-            console.log(collection_name);
-            console.log(['Removing: ', await db.collection(collection_name).countDocuments({}), ' documents from ', collection_name].join(''));
-            await db.collection(collection_name).deleteMany({});
+            await client.close();
+            res.status(200).json({success: true, collections: collection_names});
         }
-        await client.close();
-        res.status(200).json({success: true, collections: collection_names});
+    } else {
+        res.status(403).json();
     }
 });
 
@@ -110,7 +116,7 @@ app.get('/stats', async (req, res) => {
     } else {
         //create client by providing database name
         const db = client.db(mongoConfig.db_name);
-        var collection_names = [];
+        let collection_names = [];
         // const collections = await db.listCollections().toArray();
 
         await db.listCollections().forEach(collection => {
@@ -120,7 +126,7 @@ app.get('/stats', async (req, res) => {
             }
         });
 
-        var collection_stats = [];
+        let collection_stats = [];
         console.info('Collection_names:' + collection_names);
         for (const collection_index in collection_names) {
             const collection_name = collection_names[parseInt(collection_index)];
